# Add a new tile scheduler for varlen prefill to avoid launching empty work groups