@Antonyvance 2025-10-17
Can you check this PR  [547](https://github.com/intel/sycl-tla/pull/547) and redesign accordingly? @carsonwang